Move set_allocator to utils.py
